Add ETHUSDT 1h data conversion step to data consolidation process
